Add long_query_time option to mysqldump
Summary: The slow log is currently contains noise from mysqldump because their queries are usually longer running. This adds an option to mysqldump to specify long_query_time as a session variable to avoid this problem.

Specifying --long-query-time=0 for mysqldump means using the server's value instead of explicitly setting it. It does not set long_query_time to 0.

Test Plan: mysqltest.sh mysqldump-long-query-time
./mtr --parallel=16 --repeat=16 main.mysqldump-long-query-time main.mysqldump-long-query-time ...

Reviewers: ebergen, jtolmer

Differential Revision: https://reviews.facebook.net/D47499