Fixed a crash in join buffer for left join due to Field_document::is_null() accessing invalid blob data
Summary:
Join buffer is used in the join execution for this self left join. Firstly, the related table column data is loaded in the join buffer, then it iterates over the table again and for each row match the rows in the join buffer. When matching a column in a row in the join buffer, the flags will be populated in the field instance by the data in join buffer then the flags will be checked, e.g. by calling is_null() to find out if the column is null, if it is then it will be skipped otherwise the payload will be loaded from join buffer and the matching will continue. Document columns are treated as blobs here, which is ok, but Field_document implements the virtual function is_null(), in which not only the flags are checked but also the payload is checked, this is the new semantics introduced for derived document field and document path, which is fine in general, but in this particular scenario, the field instance is partially populated, i.e., the flag data is valid but the blob payload is not, this caused Field_document::is_null() returning wrong result which caused incorrect pointer increments, which caused the following operations accessing invalid memory and eventually crashed the server. To fix this issue, a new method Field_document::is_null_as_blob() is introduced to check if the document column is null as a blob, in which only the flags will be checked.

This diff also added and updated some test cases for large document column, group by and order by with document paths,etc.

Test Plan: json.type_document_path.test

Reviewers: santoshb

Subscribers:

Differential Revision: https://reviews.facebook.net/D47571