Fix lint errors in the current code.
Summary:
We are adding linting to the MyRocks code.  The linters only show warnings on changed lines,
but all errors are shown whether the line has changed or not.  This diff is to fix all the errors that
lint finds so that everybody else doesn't have to deal with them.

Many of the errors that showed up were related to the header files so I also cleaned them up, reorganizing
them so that they had a consistent order.

Differential Revision: https://reviews.facebook.net/D48021

fbshipit-source-id: 18255ac89e9