Fixed the connection throughput regression introduced in the 5.7 backport of extending protocol's OK packet
Summary: This diff fixed the connection throughput regression introduced in the 5.7 backport of extending protocol's OK packet. That backport introduced heavy session tracker initialization in THD::init() which added extra workload for the socket accept thread, this makes the accept thread accepts less number of connections. This diff uses lazy evaluation to defer the initialization to worker threads to fix the regression. Now the connection throughput back to the number before the backport. The improvement in my test was about 8.7%.

Test Plan: MTR

Reviewers: tianx, 1251333826, drilibo, santoshb, jtolmer

Reviewed By: drilibo

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D48465