Allow rocksdb_rate_limiter_bytes_per_sec to be changeable
Summary:
Currently rocksdb_rate_limiter_bytes_per_sec can only be used to set up a rate limiter during
startup.  If the bytes_per_sec value needs to change after startup, that is not possible.  Change this
to allow for this possibility.  If no rate limiter was setup during startup you won't be able to start
and if one is set up during startup you won't be able to disable it.  But if one is set up during startup
the value for the bytes_per_sync will be changeable

Test Plan: MTR with expanded rocksdb_rate_limiter_bytes_per_sec_basic.test

Reviewers: maykov, yoshinorim, hermanlee4

Subscribers: