Increase timeout in main.partition_innodb_plugin
Summary:
The main.partition_innodb_plugin test is occasionally failing due to the
table not returning any rows. A prior fix added a timeout to ensure the
data is available. Increase this timeout to a value much larger than the
min cache update to ensure the data is available on a read.

Squash with Fix flaky test main.partition_innodb_plugin

Test Plan: Run the main.partition_innodb_plugin

Reviewers: jtolmer, tianx, maykov

Subscribers: