Convert index_id into (cf_id, index_id).
Summary:
Eventually we want to be able to import tables and the index_id values from the other tables
may collide with existing index_ids.  Thus we are converting the global index identifier from a single
integer to the combination of a column family and index id.

@update-submodule: rocksdb

Test Plan: MTR

Reviewers: yoshinorim, hermanlee4, maykov

Differential Revision: https://reviews.facebook.net/D48477