Add support for READ_COMMITTED transaction mode and delay snapshots until first operation
Summary:
Before this we did not support READ_COMMITTED transaction mode.  Add support for this.  This involves creating a new snapshot for each statement that is executed so that other committed changes will be seen.  Additionally, for REPEATABLE_READ, delay grabbing a snapshot until the first RocksDB operation (Put, Get, Delete, etc.) is needed.

update-submodule: rocksdb

Differential Revision: https://reviews.facebook.net/D51993

fbshipit-source-id: 90a85c01ba7