Better error detection in innodb_stress tests
Summary:
Refactor load_generator.py so that there is a bit less code
duplication among the various worker types. Doing so allows
for consistent error detection so that main can look at the
worker to know if an unhandled expception was encountered.
If one was, fail the test instead of silently ignoring it.

Fix DefragmentWorker to work correctly with crash tests by
handling "Server has gone away" errors and trying again.

Create, examine, and drop the errors table that load_generator.py
was using. If rows get inserted into it during the test, the test
will fail with mismatched output.

Once tested and found to be stable, this commit will need to be
divided up and squashed with multiple previous commits, some from
upstream, including:
  Add basic suite of innodb stress tests
  Implement online defragmentation
  Add support for disabling defrag in stress tests

Test Plan: Jenkins

Reviewers: hermanlee4, santoshb

Differential Revision: https://reviews.facebook.net/D48753