Better error detection in innodb_stress tests
Summary:
Refactor load_generator.py so that there is a bit less code
duplication among the various worker types. Doing so allows
for consistent error detection so that main can look at the
worker to know if an unhandled expception was encountered.
If one was, fail the test instead of silently ignoring it.

Test Plan: Jenkins

Reviewers: hermanlee4, santoshb

Subscribers: