Adding ROCKSDB_DATA_DIR system variable
Summary:
The directory in which the rocksdb database is stored is now
configurable.

Test Plan:
There is one new automation test case:
sys_vars.rocksdb_data_dir_basic

The functionality can be tested by modifying mysql-test-run.pl so that
the argument can also be passed to bootstrap, then running the rocksdb.1st test
(using the new system variable) and verifying that .sst files are
created and placed in the directory specified as the argument to the
parameter.

Reviewers: maykov

Differential Revision: https://reviews.facebook.net/D47631