call ha_recover() irrespective of the state of binlog
Summary:
Crash recovery should call ha_recover() to fetch the storage engine's
binlog coordinates which is used by MySQL to fix binlog state. But, in the event
of a crash during binlog rotate where the old binlog is closed cleanly but the
new binlog is not yet opened, ha_recover() is not called.

Squash with "Crash safe slave and master in RocksDB"

Test Plan:
Added a mtr test.
mtr test fails without this fix.

Reviewers: jtolmer

Subscribers: webscalesql-eng, yoshinorim

Differential Revision: https://reviews.facebook.net/D49221