Force manual compaction through drop_index_thread to remove dropped index.
Summary:
We now use manual compactions to compact away indexes which
were dropped. This is done in the drop_index_thread. The eliminating of
rows from drop indexes is done in the compaction filter which is called
when manual compaction is done on the range covering the dropped index.
@update-submodule: rocksdb

Test Plan: rocksdb.drop_table

Reviewers: maykov, jkedgar, spetrunia, yoshinorim, hermanlee4

Reviewed By: hermanlee4

Subscribers: sdong, webscalesql-eng

Differential Revision: https://reviews.facebook.net/D49107