Using a UUID for index id to see what the effects are
Summary: Currently the index id is made up of the column family id and an incrementing integer (total of 8 bytes), but it still may not be unique when we begin support transportable tables.  This change uses a uuid (16 bytes) instead to attempt to avoid the collision on IDs.

Test Plan: MTR

Reviewers:

Subscribers: