Fix row count and add more fields for audit plugin
Summary: It seems like the audit plugin implementation MySQL is reusing
a variable that does not always reflect rows sent to client. This means that
that in many cases, it is not counting the actual rows sent to the client.
There is a more accurate count on the THD object that should be used instead.

I am ignoring the documentation so that it just returns the number of rows
returned the client:
(https://dev.mysql.com/doc/refman/5.5/en/writing-audit-plugins.html)

I am also adding new fields that will be passed into the audit plugin. This
includes, number of affected rows, the connection certificate, and query
attributes.

Query attributes are currently not implemented, but will come shortly.
Connection certificates will also come afterwards.

Test Plan: checked values manually in debugger for
select/insert/updates/deletes.

Reviewers: tianx, jtolmer

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D49329