Fix row count for audit plugin
Summary:
It seems like the audit plugin implementation MySQL is reusing a variable that does not always reflect rows sent to client. This means that that in many cases, it is not counting the actual rows sent to the client. There is a more accurate count on the THD object that should be used instead.

According to the documentation (https://dev.mysql.com/doc/refman/5.5/en/writing-audit-plugins.html), it seems like STATUS events always return 0 for rows.

Test Plan: manually with queries (SELECT/UPDATE/DELETEs)

Reviewers: tianx, jtolmer

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D49329