Fix compile on Mac Clang
Summary:
Not sure if this is a Mac thing or a Clang thing, but this
fails to build with clang 3.6 on my Mac, due to an incompatible
comparison between a char* and an int. Pretty sure it is in fact wrong
-- it looks like it wants to be checking if the first character is a
null byte.

Test Plan:
Builds on my Mac with clang 3.6 as part of HHVM. Relying on
Jenkins for the rest.

Reviewers: drilibo, tianx

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D49677