Fix main.log_state for long paths
Summary:
I saw some test failures in our CI system from main.log_state
because the path is long and the error message truncates it to 200
characters. Fix the test by adding a second value for --replace_result
that is also truncated to 200 chars.

Test Plan:
Move my local repository to a long path and then:
../tools/mysqltest.sh --nobuild main.log_state

Reviewers: mung

Reviewed By: mung

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D49683