[AsyncClient] Implement Async Handshake
Summary:
What's done:
This implements the client layer on top of viossl to enable async calls to `sslconnect`. In order to have a proper async handshake, we need
1- Send a request packet with client capabilities after reading the initial scramble
2- Do the SSL Exchange (that's on viossl layer)

This code was tangled in the the `prep_client_reply_packet` in a very synchronous way. This diff breaks it apart into smaller functions that saves the context in both ssl vio
- `SSL*` - this is to avoid the multiple calls for sslconnect to restart a new SSL struct
- SSL connect state for `send_client_reply_packet_nonblocking` to know in which write action it's doing.

There were some other weird things going on. I went over several solutions until this one (closest to the original code) worked for most cases. Since we need to read the initial packet before we start writing to net, this SSL connect needs to be down the chain, not in the main state machine.

In addition to that, `sslconnect` now returns more signs to the client
about the actions it should take, if the SSL_connect requires READ or
WRITE. Then we set this value to `async_blocking_state`. This value is
used to mark to libevent what event we want to be woken up for.

Current flow:
- Right before we start the authentication part of the protocol, the client code and other codepaths call `send_client_reply_packet` that calls `prep_client_reply_packet`. So I removed the sync code from inside of `prep_client_reply_packet`.
- Now we have `send_client_reply_packet` and `send_client_reply_packet_nonblocking` each doing `run_ssl_request` and `run_ssl_connect` and finally sending the first actual authentication packet.

What's coming up:
- Clean up, I have a lot of code out of places since I moved code around a lot.
- Fix the white spaces :P
- I'm hard coding the size of the first request packet, I need to make it prettier
- See if all tests pass, small note here, `--ssl` was disabled --async-client, so I fixed it and during the tests I added ASSERTS and checked traces to make sure I was seeing the tests doing the right thing

Test Plan: Making sure the nonblocking calls for SSL were actually being hit during tests.

Reviewers: chip, 1251333826

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D50223