Another fix for innodb.innodb_alter_table_defragment
Summary:
Killing a running defrag command was causing problems with later parts
of the test because defrag structures are updated asychronously by the
defrag thread, not synchronously on the thread that gets back the error.
I didn't see any status variables I could use with wait_for_condition.inc,
so I moved the restart_server that was put in place previously to the point
in the test where it's needed.

Test Plan: Jenkins

Reviewers: kradhakrishnan, tianx

Reviewed By: tianx

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D49665