Fix innodb_doublewrite on Ubuntu
Summary:
The innodb_doublewrite python script was failing to terminate the mysqld process, which was causing problems for the test. The reason is because shell=True was set in subprocess.Popen, the python script does not directly spawn the mysqld process. It spawns a shell process, which spawns the mysqld process. When python tries to terminate its child, it is actually terminating the shell, and not the mysqld process. The fix is to set shell=False, since there's no reason for it to be true.

The reason we don't hit this test is because we're using bash as the default shell, whereas Ubuntu uses dash. Whereas bash will replace its current process with the new process to be spawned, it appears that dash will just fork and create a child process. This means that we end up terminating the correct process when using bash. On Ubuntu, the problem also goes away if you set the default shell to bash.

Test Plan: ./mtr innodb_doublewrite on Ubuntu

Reviewers: jtolmer

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D50535