Fixing assertion failure "array->n_elems < array->max_elems" in stress tests
Summary:
Big stress tests sometimes fail with this assertion in debug build, due to
running out of the slots reserved for each thread in sync level array. The size
is hardcoded to SYNC_THREAD_N_LEVELS, and never increases. This patch expands
the sync level array when all slots are used up.

Test Plan:
Manually set SYNC_THREAD_N_LEVELS to a small number (so the sync level array is
very small). Without the fix, stress tests failed with the assertion. With the
fix, tests succeeded with the fix.

Reviewers: jtolmer

Subscribers: