[RFC] [Async Mysql] Forward the SSL socket action back to client
Summary:
This allows the client to know what event VIO_SOCKET wants so we can
properly schedule the event in the IO Thread.

`async_blocking_state` needs to be set correctly so all the way up in:
https://phabricator.fb.com/diffusion/FBCODE/browse/master/squangle/mysql_client/Operation.cpp;de4e7e56fdaaa4b65655fff2cadbca6fbaceec80$52-80
this is scheduled.

Thus, we get the right amount of events, less loops.

Opening a new diff because I'm not happy with how I'm passing the action
back to client, changing what `sslconnect` returns is weird, but we are
the only users other than tests. The other option would be to add
another parameter to set the `vio_action`.

Test Plan: Unit tests and manual testing with `mysql_test_tool`

Reviewers: anca, chip

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D50655