Fix an assert when using expand_fast_index_creation
Summary:
If expand_fast_index_creation is on and a secondary index in the altered table references the last field in the table an assertion was being triggered.  This was related to MySQL sometimes storing field numbers in the key information as zero based values (0 to N-1) and sometimes as one based values (1 to N).  One of the functions being called by the code that was added for expand_fast_index_creation was mistakenly passing one-based values to a function that was expecting zero-based values.

Reported at https://github.com/facebook/mysql-5.6/issues/99

Squash with 2e2ea2fa863885111ea2626a887e2067096c4f10
(Port Fast Index Creation to 5.6)

Test Plan: MTR

Reviewers: hermanlee4, jtolmer, spetrunia, santoshb

Reviewed By: santoshb

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D50715