Stabilize the innodb.innodb_histogram_show_innodb_status test that is occasionally flakey in parallel tests.
Summary:
The innodb.innodb_histogram_show_innodb_status will sometimes fail, particularly when parallel test are being run. (https://our.intern.facebook.com/intern/tasks/?t=8998622).  The test queries the total of some counters and then queries the individual values and assumes the total and the sum of the individual values will be equal (or off by at most one).  This is not always the case.  While crushing the test (16 parallel executions repeated 16 times) I was able to see a difference of up to 7.  I changed the test to allow a difference of up to 7.  We may need to increase it in the future.

Squash with 5ee7572 Add histograms for tracking I/O and per user latencies

Test Plan: MTR

Reviewers: jtolmer, hermanlee4

Reviewed By: hermanlee4

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D50751