LOG events for connection are missing the query field in audit plugin
Summary:
The LOG events in the audit plugins used to contain information about the user/ip/db in the query field. This was broken when Oracle fixed some other bug (38dcb9).

This adds the fields back in.

Test Plan:
Break on mysql_audit_general_log in gdb and connect to the server. Verify that the field was empty before and contains information after this change.

Reviewers: jtolmer drilibo

Subscribers: