LOG events for connection are missing the query field in audit plugin
Summary:
The LOG events in the audit plugins used to contain information about the user/ip/db in the query field. This was broken when Oracle fixed some other bug (38dcb9).

This adds the fields back in.

Test Plan: Break on mysql_audit_general_log in gdb and connect to the server. Verify that the field was empty before and contains information after this change.

Reviewers: drilibo, jtolmer

Reviewed By: jtolmer

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D50817