[Client] Only assign insert_id and affected_rows on query result
Summary:
As of the backport of
`WL#4797 Extending protocol's OK packet`
the OK Packet got read in a separate function, making every time we
parse it, we assign the values of `insert_id` and `affected_rows`. This
is a simple fix to only update `MYSQL*` when it's a result of a query.

Test Plan:  unit tests

Reviewers: pengt, anca, tianx