Temporary fix to prevent dropping tables from crashing the server on startup
Summary:
Due to a race condition, it is possible for rocksdb to start up
compaction and call into MyRocks's compaction filter code before the
data dictionary is fully initialize. The proper fix is to add a wait to
rocksdb. This is a temporary fix that forces the compactions threads
to wait until data dictionary is initialized.

Test Plan: run unit tests

Reviewers: spetrunia, yoshinorim, jkedgar

Subscribers: