Add query attributes
Summary:
Currently, queries sent over the network are encoded as follows:
  total packet length
  COM_QUERY
  query string

The length of the query string is implicit from and calculated from the packet length. This change adds a new command called COM_QUERY_ATTRS with the following structure
  total packet length
  COM_QUERY_ATTRS
  total attributes length
  attributes string
  query string

Note that query string length is still implicit.

The attributes are stored on the THD object. This information will be sent to the audit plugin.

Test Plan: mysqltest.sh query_attrs

Reviewers: jtolmer, anca, jkedgar

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D51105