Add query attributes
Summary:
Currently, queries sent over the network are encoded as follows:
total packet length
COM_QUERY
query string

The length of the query string is implicit from and calculated from the packet length. This changes the format to the following:
total packet length
COM_QUERY
total attributes length
attributes string
query string

Note that query string length is still implicit. As well, this new format is only expected if the server turns on the `query_attrs` global variable, and the client also sets `CLIENT_QUERY_ATTRS`. Otherwise, the old behaviour is maintained.

The attributes are stored on the THD object. This information will be sent to the audit plugin. This part is not implemented yet.

Test Plan:
mysqltest.sh query_attrs

Testing is done by outputting the query attributes in the error log if they are received, and then checking the error log. Note that we are implicitly testing all 4 combinations possible.

1. Old client with old server (all our existing tests that use the default connection)
2. New client with old server (since `CLIENT_QUERY_ATTRS` is only set for `connect_n_handle_errors` and not `safe_connect`, this case arises whenever we issue the `connect` command in mysqltest)
3. New client with new server (`query_attrs.test`)
4. Old client with new server (`query_attrs.test`)

Reviewers: jtolmer, drilibo, anca

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D51105