disable compaction until after structures needed for compaction filter are initialized
Summary:
This is the myrocks side changes for delaying compaction until the appropriate
structures needed are initialized.

This goes along with rocksdb changes that expose
EnableAutoCompactions in the public api.

Test Plan:
was able to reproduce the seg fault mentioned in https://github.com/facebook/mysql-5.6/issues/73
using
../tools/mysqltest.sh --mem --force rocksdb.drop_table

method was to manually sleep the thread after DB::Open but before TransactionDB ptr was
assigned in transaction_db_impl.cc:
  DB::Open(db_options, dbname, column_families_copy, handles, &db);
  clock_t goal = (60000 * 10) + clock();
  while (goal > clock());
  ...dbptr(aka rdb) gets assigned below

verified my changes fixed the issue.

also wrote a  test to ensure the compaction was turning on and off, but it has
a flakiness factor to it due to the way log flushes, can upload if needed

Reviewers: hermanlee4

Reviewed By: hermanlee4

Subscribers: yoshinorim, jkedgar, webscalesql-eng

Differential Revision: https://reviews.facebook.net/D51153