Implemented a Monotonic Timestamp for Getting Sec Behind Master
Summary:
The background is,  if we do things like replay Binlogs on a master, second behind master will be flapping back and forth between 'current' and time delta against statements with 'SET TIMESTAMP' on the master. This is because statements on the master can SET TIMESTAMP to a non-linear value (an arbitrary value that is crazily small or crazily large).  This essentially causes the second behind master of the slave to be flapping back and forth.

This diff addresses this issue by keeping the last_master_timestamp monotonic during the entire run without resetting it to zero. We added an additional flag in relay log info to represent that the slave has caught up (originally represented by last_master_timestamp == 0).  Since this diff keeps the last_master_timestamp monotonic, this flag representing that the slave has caught is needed. This diff added necessary logics to update the flag and also use this flag to update the second behind master.

Essentially, for both parallel slave worker and non-parallel slave worker scenarios, the last_master_timestamp is updated only when the new timestamp from the new event is bigger than the last_master_timestamp.  We set the maximum to be current slave time to prevent last_master_timestamp from going crazily large. This basically prevent the second behind master from jumping back and forth, which is caused by inappropriately "SET TIMESTAMP" from the master.

Also changed rpl.rpl_seconds_behind_master.test that is affected by this diff. The is because the test has decreased the master events' timestamp by 3600 in the middle and asserted that the second behind master should be greater than 3600. This is not supposed to happen with this diff since this is the spike that we are trying to solve in this diff. I have changed the test file's description and assertion. Now we assert that the second behind master should be smaller than 3600 instead.

Test Plan:
```
mysqltest.sh rpl.rpl_slave_monotonic_timestamp
mysqltest.sh rpl.rpl_seconds_behind_master

```
The test covers both parallel slave worker and non-parallel slave worker scenarios due to the different code paths.  The test launches a heavy amount of queries on the master with flapping timestamps and keep polling the "second behind master" on the slave to assert the second behind master is smaller than a threshold (Smallest timestamp is set around 1000000000 on the master, so the threshold should be smaller than 400000000).

Without this code changes in this diff, the problematic scenarios can be reproduced by the test file for both MTS and Non-MTS cases.

Reviewers: santoshb

Reviewed By: santoshb

Subscribers: domas, webscalesql-eng

Differential Revision: https://reviews.facebook.net/D51207