fix failure on rpl.rpl_thread_name test
Summary:
The rpl_thread_name test was failing under certain conditions because the binlog
dump thread is renamed from a connection thread initially, but not changed back
to the regular connection thread name afterwards if it used as something other
than a binglogd thread in the future.

To fix this we, check for the type of incoming command, if it isnt binlog dump
command then we know to name the thread to a regular connection thread.

Squash with D50523

Test Plan:
Run
../tools/mysqltest.sh --clean --perfschema rpl.rpl_thread_name --parallel=16 --repeat=16
observe that it now passes.

Reviewers: tengli, jtolmer, jkedgar

Reviewed By: jkedgar

Subscribers: santoshb, jkedgar, webscalesql-eng

Differential Revision: https://reviews.facebook.net/D51237