fix failure on rpl.rpl_thread_name test
Summary:
The rpl_thread_name test was failing under certain conditions because the binlog
dump thread is renamed from a connection thread initially, but not changed back
to the regular connection thread name afterwards if it used as something other
than a binglogd thread in the future.

To fix this we set the name back to regular connection thread on thread end.

Squash with D50523

Test Plan:
Run
../tools/mysqltest.sh --clean --perfschema rpl.rpl_thread_name --parallel=16 --repeat=16
observe that it now passes.

Reviewers: jtolmer, tengli

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D51237